"Hay"-now, this commit adds items to the plugin. What a luxury. Every items is defined by a name and an ItemStack.
This config is used to reference items in other configs like the ship config. Right now this is not used for blocks but this functionality will be added later.

The Items are saved in a new config called items.yml.
Items are stored at runtime in the BattleshipItems class.
Items can be retrieved using the getItem method in the BattleshipItems class.
For the path every . is replace with a § in the config (Might need a better way for this).
As for debugging there a two new subcommands for the mdebug command:
 - /mdebug addItem <path> //The item in your hand will be added
 - /mdebug saveItemConfig