Ouuuuuppppsss... This commit fixes the latest introduced bug and also adds a new Gui.
This commit adds some sort of Gui system to the plugin. It also adds the first gui, the Map selection gui.
Now if you challenge a player the gui pops up and wants you to click on the map. After that the message is sent to the opponent with the map name. The game now also knows about the selected map now :D